test: isolate the sweep tests from every other save in the process
The ubuntu leg failed on a_leaked_temp_is_swept_for_a_long_non_ascii_
destination_name while macOS and Windows passed, which looks like a
platform bug and is not one. The sweep is best-effort by design:
claim_first_sweep takes SWEPT with try_lock, not lock, so that a fork
landing mid-hold cannot leave the child with a locked mutex and hang its
first write. Declining just skips an opportunistic sweep.

That makes any concurrent save able to turn a sweep assertion into a
failure for a reason unrelated to the sweep's logic. io_hardening runs
~40 tests in parallel and many of them save, so this is a live race, not
a Linux quirk: running that suite 40 times here reproduced a spurious
failure, and re-running it surfaced a *different* sweep test
(the_sweep_spares_a_shorter_destinations_temp_that_looks_truncated),
which is the signature of contention rather than of any one test.

SWEPT is a process-global static and each integration test file is its
own process, so move the five sweep tests into their own binary. The
only saves in that process are these tests, and a file-local mutex keeps
even those from overlapping. The product code is untouched — the tests
stop competing for the lock rather than the lock changing.

Verified: 60 consecutive runs of both binaries, zero failures, against
roughly 1 in 40 before.
